MACHINE LEARNING/STREAMLIT/case-2/requirements.txt
streamlit 
torch 
torchvision 
timm 
pillow
requests